pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "npuwattch"
dynamic = ["version"]
description = "NPUWattch"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch",
  "PyYAML>=6.0"
]

[project.scripts]
npuwattch = "npuwattch.npuwattch_console:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "npuwattch._version.__version__"}
